services/ms-auth-core/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
PyJWT[crypto]>=2.8,<3.0
prometheus-client>=0.20,<1.0
